docs: fix build script and add README

+++ paper/build.sh
@@ -1,8 +1,8 @@
 #!/bin/bash
 
 rm -r _minted-paper paper.aux paper.bbl paper.blg paper.log paper.out paper.synctex.gz build.log >> /dev/null 2>&1
-docker stop latex
-docker rm latex
+docker stop latex >> /dev/null 2>&1
+docker rm latex >> /dev/null 2>&1
 set -x
 cp biblio/references.bib paper.bib
 
@@ -18,15 +18,15 @@ docker run --rm \
      pdflatex -synctex=1 -interaction=nonstopmode -shell-escape paper; \
      pdflatex -synctex=1 -interaction=nonstopmode -shell-escape paper;" &> build.log
 
-#ERR=$?
-#if [ $ERR != "0" ] ; then
-#  set +x
-#  echo "ERROR: $ERR"
-#  cat build.log
-#  exit 1
-#fi
-
+ERR=$?
 set +x
+if [ $ERR != "0" ] ; then
+  set +x
+  echo "ERROR: $ERR"
+  cat build.log
+  exit 1
+fi
+
 rm -r _minted-paper paper.aux paper.bbl paper.blg paper.log paper.out paper.synctex.gz build.log >> /dev/null 2>&1
 
 echo "SUCCESS"
